backend/requirements.txt
# FastAPI and dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pytest

# LLM APIs
openai>=1.72.0
google-genai>=0.1.0

# Utilities
httpx>=0.26.0
python-multipart>=0.0.9
Pillow>=10.0.0  # For image processing
